@@ -3,7 +3,7 @@ module Spree::MailChimp
     class << self
       def instance
         return nil unless ActiveRecord::Base.connection.tables.include?('configurations')
-        MailChimpConfiguration.find_or_create_by_name("MailChimp configuration")
+        MailChimpConfiguration.find_or_create_by_name('MailChimp configuration')
       end
     end
   end
