@@ -1,12 +1,18 @@
 module SpreeMailChimp
   class Engine < Rails::Engine
+    require 'spree/core'
+    isolate_namespace Spree
     engine_name 'spree_mail_chimp'
 
     config.autoload_paths += %W(#{config.root}/lib)
 
+    config.generators do |g|
+      g.test_framework :rspec
+    end
+
     def self.activate
-      Dir.glob(File.join(File.dirname(__FILE__), "../../app/**/*_decorator*.rb")) do |c|
-        Rails.application.config.cache_classes ? require(c) : load(c)
+      Dir.glob(File.join(File.dirname(__FILE__), '../../app/**/*_decorator*.rb')) do |c|
+        Rails.configuration.cache_classes ? require(c) : load(c)
       end
     end
 
